"Maybe if I have two divs full of content?"
This is an idea born out of Friday-Night desperation: that maybe if we
just have two copies of the data, but in different hide-able divs, with
the orientations we need...   Yeah.  I'm keeping it here for now in case
it's so crazy it might work, otherwise I expect this branch to not go
anywhere as I focus on just making something that looks good on the
iPad.